requirements.txt
torch>=1.9.0
torchvision>=0.10.0
matplotlib>=3.4.0
numpy>=1.20.0
Pillow>=8.2.0
streamlit>=1.10.0
pytest>=6.2.5
